Moved TesouroFragment and TesouroAdapter to tesouro package
