+++ app/src/main/java/sistemas/puc/com/finantialapp/tesouro/TesouroCursorAdapter.java
@@ -1,4 +1,4 @@
-package sistemas.puc.com.finantialapp.adapters;
+package sistemas.puc.com.finantialapp.tesouro;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -10,6 +10,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import sistemas.puc.com.finantialapp.R;
+import sistemas.puc.com.finantialapp.adapters.AbstractCursorAdapter;
 import sistemas.puc.com.finantialapp.data.FinantialContract.TesouroEntry;
 import sistemas.puc.com.finantialapp.util.Util;
