+++ app/src/main/res/layout/fragment_tesouro.xml
@@ -5,4 +5,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="0dp"
-    tools:context=".TesouroFragment" />
+    tools:context=".tesouro.TesouroFragment" />
